// src/app/shared/model/CustCompanyLegalDocObj.Model.ts
export class CustCompanyLegalDocObj {    
    CustCompanyLegalDocId: any;
    CustCompanyId: any;
    MrLegalDocTypeCode: any;
    DocNo: any;
    DocDt: any;
    DocExpiredDt: any;
    DocNotes: any;
    NotaryName: any;
    NotaryLocation: any;
    RowVersion: any;
    constructor(){this.CustCompanyLegalDocId = 0; this.RowVersion = "";}
}